move recordToDataCite() to its own file
